further fixture modifications
