+++ tests/test_DataFrameWrapper.py
@@ -6,6 +6,7 @@ from typing import Any, Dict, Iterable, Optional, Sequence, Union
 import pytest
 from chispa.dataframe_comparer import assert_df_equality
 from dask.array import concatenate, ones, zeros
+from fixtures.fixtures_metadata import *
 from numpy import arange
 from pyspark.sql import DataFrame as SparkDataFrame
 from pyspark.sql import SparkSession
@@ -25,7 +26,6 @@ from xarray.testing import assert_identical
 from sds_data_model.constants import BNG_XMAX, BNG_XMIN, BNG_YMAX, BNG_YMIN, CELL_SIZE
 from sds_data_model.dataframe import DataFrameWrapper
 from sds_data_model.metadata import Metadata
-from sds_data_model.tests.fixtures.test_metadata import expected_metadata
 
 
 @fixture
